Refactor HTTP response handling into focused modules
